refactor(frontend): normalize feature support structure
