+++ frontend/src/lib/planner/creditParsing.ts
