+++ frontend/src/providers/AppProviders.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { AppThemeProvider } from "@/components/theme/ThemeProvider";
-import { ScheduleProvider } from "@/context/schedule-context";
+import { ScheduleProvider } from "@/context/schedule/schedule-context";
 
 export function AppProviders({ children }: { children: React.ReactNode }) {
   return (
